Restruct for submodule.
